kinda getting rotation working
but the event listeners aren't really set up well, so behaviours are unexpected. mousemove and mouseup events aren't being fired unless the pointer is within the bounds of the shape being rotated, which leads to screwy behaviour. Need to have some window-scoped event listeners and global scoped 'dragging'  flag to correct this.